+++ processor/src/test/fixtures/bad_input/com/example/injector/Jsr330ScopedInjectorModel.java
@@ -1,0 +1,12 @@
+package com.example.injector;
+
+import javax.inject.Singleton;
+import sting.Injector;
+
+@SuppressWarnings( "CdiManagedBeanInconsistencyInspection" )
+@Singleton
+@Injector
+public interface Jsr330ScopedInjectorModel
+{
+  String getMyThing();
+}