+++ processor/src/test/fixtures/bad_input/com/example/injector/outputs/Jsr330ScopedOutputModel.java
@@ -1,0 +1,11 @@
+package com.example.injector.outputs;
+
+import javax.inject.Singleton;
+import sting.Injector;
+
+@Injector
+public interface Jsr330ScopedOutputModel
+{
+  @Singleton
+  String getMyThing();
+}