+++ processor/src/test/java/sting/processor/StingProcessorTest.java
@@ -434,6 +434,8 @@ public final class StingProcessorTest
 
         new Object[]{ "com.example.injector.ClassInjector", "@Injector target must be an interface" },
         new Object[]{ "com.example.injector.EnumInjector", "@Injector target must be an interface" },
+        new Object[]{ "com.example.injector.Jsr330ScopedInjectorModel",
+                      "@Injector target must not be annotated with an annotation that is annotated with the javax.inject.Scope annotation such as [@javax.inject.Singleton]" },
         new Object[]{ "com.example.injector.MultipleCandidatesForSingularDependencyModel",
                       "@Injector target must not contain a non-collection dependency [java.lang.Runnable] that can be satisfied by multiple nodes.\n" +
                       "  Dependency Path:\n" +
@@ -518,6 +520,8 @@ public final class StingProcessorTest
 
         new Object[]{ "com.example.injector.outputs.ArrayTypeOutputModel",
                       "@Injector target must not contain a method with a return type that contains an array type" },
+        new Object[]{ "com.example.injector.outputs.Jsr330ScopedOutputModel",
+                      "@Injector target must not contain a method that is annotated with an annotation that is annotated with the javax.inject.Scope annotation such as [@javax.inject.Singleton]" },
         new Object[]{ "com.example.injector.outputs.NullableCollectionOutputModel",
                       "@Injector target must not contain a method annotated with @Nullable that is not an instance dependency kind" },
         new Object[]{ "com.example.injector.outputs.NullableSupplierCollectionOutputModel",